+++ app/(admin)/admin/application/[id]/styled/index.ts
