+++ app/(admin)/admin/styled/index.ts
