+++ app/(user)/apply/styled/index.ts
