+++ app/(user)/certificate/styled/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { styled } from 'styled-components';
 
 export const CertificateContainer = styled.main`
