+++ app/lib/index.tsx
@@ -1,3 +1,3 @@
 export * from './Provider';
 export * from './authOptions';
-export * from './fonts';
+export * from './Fonts';
