+++ components/Admin/ApplicationDetail/Application/Application.styled.ts
