+++ components/Admin/ApplicationDetail/Application/ApplicationStatus/ApplicationStatus.styled.ts
