--- components/Admin/ApplicationList/ApplicationDetail/index.tsx
@@ -1,3 +1,0 @@
-import ApplicationDetail from './ApplicationDetail';
-
-export default ApplicationDetail;
