--- components/Admin/ApplicationList/ListComponent/Status/index.tsx
@@ -1,3 +1,0 @@
-import Status from './Status';
-
-export default Status;
