--- components/Admin/ApplicationList/ListComponent/index.tsx
@@ -1,3 +1,0 @@
-import ListComponent from './ListComponent';
-
-export default ListComponent;
