--- components/Apply/index.tsx
@@ -1,3 +1,0 @@
-import Apply from './Apply';
-
-export default Apply;
