--- components/Certificate/Certificate.tsx
@@ -1,180 +1,0 @@
-'use client';
-
-import Nav from '@/components/Nav';
-import { MAIN_COLOR } from '@/constants';
-import * as S from './Certificate.styled';
-
-const Certificate = () => {
-  const color = MAIN_COLOR;
-
-  return (
-    <S.CertificateContainer>
-      <Nav color={color} />
-      <S.ContentContainer>
-        <S.HeadStyle>Certificate</S.HeadStyle>
-        <S.PrivacyContainer>
-          <S.MainTitle>개인정보 처리방침</S.MainTitle>
-          <S.Text>
-            Leets(https://leets.land) 이하 &apos;Leets&apos;은(는) 「개인정보 보호법」 제30조에 따라 정보주체의
-            개인정보를 보호하고 이와 관련한 고충을 신속하고 원활하게 처리할 수 있도록 하기 위하여 다음과 같이 개인정보
-            처리방침을 수립·공개합니다.
-          </S.Text>
-          <S.Text>○ 이 개인정보처리방침은 2023년 8월 18일부터 적용됩니다.</S.Text>
-          <S.TitleContainer>제1조(개인정보의 처리 목적)</S.TitleContainer>
-          <S.Text>
-            Leets은(는) 다음의 목적을 위하여 개인정보를 처리합니다.
-            <br />
-            처리하고 있는 개인정보는 다음의 목적 이외의 용도로는 이용되지 않으며 이용 목적이 변경되는 경우에는
-            「개인정보 보호법」 제18조에 따라 별도의 동의를 받는 등 필요한 조치를 이행할 예정입니다.
-            <br />
-            1. 홈페이지 회원가입 및 관리 회원 가입의사 확인, 회원제 서비스 제공에 따른 본인 식별·인증, 회원자격
-            유지·관리, 각종 고지·통지 목적으로 개인정보를 처리합니다.
-            <br />
-            2. 재화 또는 서비스 제공
-            <br />
-            서비스 제공을 목적으로 개인정보를 처리합니다.
-          </S.Text>
-          <S.TitleContainer>제2조(개인정보의 처리 및 보유 기간)</S.TitleContainer>
-          <S.Text>
-            ① Leets은(는) 법령에 따른 개인정보 보유·이용기간 또는 정보주체로부터 개인정보를 수집 시에 동의받은 개인정보
-            보유·이용기간 내에서 개인정보를 처리·보유합니다.
-            <br />
-            ② 각각의 개인정보 처리 및 보유 기간은 다음과 같습니다.
-            <br />
-            1. 홈페이지 회원가입 및 관리 홈페이지 회원가입 및 관리와 관련한 개인정보는 수집.이용에 관한
-            동의일로부터(3년)까지 위 이용목적을 위하여 보유 및 이용됩니다.
-            <br />
-            보유근거 : Leets신규 회원 모집시 결과 안내
-            <br />
-            관련법령 : 신용정보의 수집/처리 및 이용 등에 관한 기록 : 3년
-            <br />
-          </S.Text>
-          <S.TitleContainer>제3조(개인정보의 제3자 제공)</S.TitleContainer>
-          <S.Text>
-            ① Leets은(는) 개인정보를 제1조(개인정보의 처리 목적)에서 명시한 범위 내에서만 처리하며, 정보주체의 동의,
-            법률의 특별한 규정 등 「개인정보 보호법」 제17조 및 제18조에 해당하는 경우에만 개인정보를 제3자에게
-            제공합니다.
-            <br />② Leets은(는) 다음과 같이 개인정보를 제3자에게 제공하고 있습니다.
-          </S.Text>
-          <S.TitleContainer>제4조(개인정보처리 위탁)</S.TitleContainer>
-          <S.Text>
-            ① Leets은(는) 원활한 개인정보 업무처리를 위하여 다음과 같이 개인정보 처리업무를 위탁하고 있습니다.
-            <br />
-            ② Leets은(는) 위탁계약 체결시 「개인정보 보호법」 제26조에 따라 위탁업무 수행목적 외 개인정보 처리금지,
-            기술적․관리적 보호조치, 재위탁 제한, 수탁자에 대한 관리․감독, 손해배상 등 책임에 관한 사항을 계약서 등
-            문서에 명시하고, 수탁자가 개인정보를 안전하게 처리하는지를 감독하고 있습니다.
-            <br />③ 위탁업무의 내용이나 수탁자가 변경될 경우에는 지체없이 본 개인정보 처리방침을 통하여 공개하도록
-            하겠습니다.
-          </S.Text>
-          <S.TitleContainer>제5조(정보주체와 법정대리인의 권리·의무 및 그 행사방법)</S.TitleContainer>
-          <S.Text>
-            ① 정보주체는 Leets에 대해 언제든지 개인정보 열람·정정·삭제·처리정지 요구 등의 권리를 행사할 수 있습니다.
-            <br />
-            ② 제1항에 따른 권리 행사는 Leets에 대해 「개인정보 보호법」 시행령 제41조제1항에 따라 전자우편 등을 통하여
-            하실 수 있으며 Leets은 이에 대해 지체 없이 조치하겠습니다.
-            <br />
-            ③ 제1항에 따른 권리 행사는 정보주체의 법정대리인이나 위임을 받은 자 등 대리인을 통하여 하실 수 있습니다.이
-            경우 “개인정보 처리 방법에 관한 고시(제2020-7호)” 별지 제11호 서식에 따른 위임장을 제출하셔야 합니다.
-            <br />
-            ④ 개인정보 열람 및 처리정지 요구는 「개인정보 보호법」 제35조 제4항, 제37조 제2항에 의하여 정보주체의 권리가
-            제한 될 수 있습니다.
-            <br />
-            ⑤ 개인정보의 정정 및 삭제 요구는 다른 법령에서 그 개인정보가 수집 대상으로 명시되어 있는 경우에는 그 삭제를
-            요구할 수 없습니다.
-            <br />⑥ Leets는 정보주체 권리에 따른 열람의 요구, 정정·삭제의 요구, 처리정지의 요구 시 열람 등 요구를 한
-            자가 본인이거나 정당한 대리인인지를 확인합니다.
-          </S.Text>
-          <S.TitleContainer>제6조(처리하는 개인정보의 항목 작성)</S.TitleContainer>
-          <S.Text>
-            ① Leets은(는) 다음의 개인정보 항목을 처리하고 있습니다.
-            <br />
-            • 홈페이지 회원가입 및 관리
-            <br />
-            • 필수항목 : 이메일, 휴대전화번호, 생년월일, 이름, 학력, 서비스 이용 기록, 접속 로그, 토큰, 접속 IP 정보
-            <br />• 선택항목 : 개인정보 동의
-          </S.Text>
-          <S.TitleContainer>제7조(개인정보의 파기)</S.TitleContainer>
-          <S.Text>
-            ① Leets은(는) 개인정보 보유기간의 경과, 처리목적 달성 등 개인정보가 불필요하게 되었을 때에는 지체없이 해당
-            개인정보를 파기합니다.
-            <br />② 개인정보 파기의 절차 및 방법은 다음과 같습니다.
-            <br />
-            파기절차 Leets은(는) 파기 사유가 발생한 개인정보를 선정하고, Leets의 개인정보 보호책임자의 승인을 받아
-            개인정보를 파기합니다. 파기방법 전자적 파일 형태의 정보는 기록을 재생할 수 없는 기술적 방법을 사용합니다.
-          </S.Text>
-          <S.TitleContainer>제8조(개인정보 자동 수집 장치의 설치•운영 및 거부에 관한 사항)</S.TitleContainer>
-          <S.Text>
-            ① Leets은 이용자에게 개별적인 맞춤서비스를 제공하기 위해 이용정보를 저장하고 수시로 불러오는 ‘토큰(token)’를
-            사용합니다.
-            <br />
-            ② 토큰는 웹사이트를 운영하는데 이용되는 서버(https)가 이용자의 컴퓨터 브라우저에게 보내는 소량의 정보이며
-            이용자들의 PC 컴퓨터내의 하드디스크에 저장되기도 합니다.
-            <br />
-            가. 토큰의 사용 목적 : 이용자가 방문한 각 서비스와 웹 사이트들에 대한 방문 및 이용형태, 인기 검색어,
-            보안접속 여부, 등을 파악하여 이용자에게 최적화된 정보 제공을 위해 사용됩니다.
-            <br />
-            나. 토큰의 설치•운영 및 거부 : 웹브라우저 상단의 도구 &gt;인터넷 옵션 &gt;개인정보 메뉴의 옵션 설정을 통해
-            토큰 저장을 거부 할 수 있습니다.다. 토큰 저장을 거부할 경우 맞춤형 서비스 이용에 어려움이 발생할 수
-            있습니다.
-          </S.Text>
-          <S.TitleContainer> 제9조 (개인정보 보호책임자) </S.TitleContainer>
-          <S.Text>
-            ① Leets은(는) 개인정보 처리에 관한 업무를 총괄해서 책임지고, 개인정보 처리와 관련한 정보주체의 불만처리 및
-            피해구제 등을 위하여 아래와 같이 개인정보 보호책임자를 지정하고 있습니다.
-            <br />
-            ② 정보주체께서는 Leets의 서비스(또는 사업)을 이용하시면서 발생한 모든 개인정보 보호 관련 문의, 불만처리,
-            피해구제 등에 관한 사항을 개인정보 보호책임자 및 담당부서로 문의하실 수 있습니다.
-            <br />
-            Leets은(는) 정보주체의 문의에 대해 지체 없이 답변 및 처리해드릴 것입니다.
-            <br />
-            ▶ 개인정보 보호책임자
-            <br />• 성명 : 김성민
-            <br />• 직책 : 회장
-            <br />• 직급 : 회장
-            <br />
-            <br />▶ 개인정보 보호 담당부서
-            <br />• 부서명 : 전략기획부
-            <br />• 담당자 : 나아연
-            <br />
-            • 연락처 : leetsland@gmail.com
-            <br />
-            ※ 개인정보 보호 담당부서로 연결됩니다.
-            <br />
-          </S.Text>
-          <S.TitleContainer>제10조(개인정보 열람청구)</S.TitleContainer>
-          <S.Text>
-            정보주체는 ｢개인정보 보호법｣ 제35조에 따른 개인정보의 열람 청구를 아래의 부서에 할 수 있습니다. <br />
-            Leets은(는) 정보주체의 개인정보 열람청구가 신속하게 처리되도록 노력하겠습니다.
-            <br />▶ 개인정보 열람청구 접수·처리 부서
-            <br />
-            • 부서명 : 전략기획부
-            <br />
-            • 담당자 : 나아연
-            <br />• 연락처 : leetsland@gmail.com
-          </S.Text>
-          <S.TitleContainer>제11조(권익침해 구제방법)</S.TitleContainer>
-          <S.Text>
-            정보주체는 개인정보침해로 인한 구제를 받기 위하여 개인정보분쟁조정위원회, 한국인터넷진흥원
-            개인정보침해신고센터 등에 분쟁해결이나 상담 등을 신청할 수 있습니다. 이 밖에 기타 개인정보침해의 신고,
-            상담에 대하여는 아래의 기관에 문의하시기 바랍니다.
-            <br />
-            1. 개인정보분쟁조정위원회 : (국번없이) 1833-6972 (www.kopico.go.kr)
-            <br />
-            2. 개인정보침해신고센터 : (국번없이) 118 (privacy.kisa.or.kr) <br />
-            3. 대검찰청 : (국번없이) 1301 (www.spo.go.kr)
-            <br />
-            4. 경찰청 : (국번없이) 182 (ecrm.cyber.go.kr) <br />
-            「개인정보보호법」제35조(개인정보의 열람), 제36조(개인정보의 정정·삭제), 제37조(개인정보의 처리정지 등)의
-            규정에 의한 요구에 대 하여 공공기관의 장이 행한 처분 또는 부작위로 인하여 권리 또는 이익의 침해를 받은 자는
-            행정심판법이 정하는 바에 따라 행정심판을 청구할 수 있습니다.※ 행정심판에 대해 자세한 사항은
-            중앙행정심판위원회(www.simpan.go.kr) 홈페이지를 참고하시기 바랍니다.
-          </S.Text>
-          <S.TitleContainer>제12조(개인정보 처리방침 변경)</S.TitleContainer>
-          <S.Text>① 이 개인정보처리방침은 2023년 8월 18일부터 적용됩니다.</S.Text>
-        </S.PrivacyContainer>
-      </S.ContentContainer>
-    </S.CertificateContainer>
-  );
-};
-
-export default Certificate;
