--- components/Certificate/index.tsx
@@ -1,3 +1,0 @@
-import Certificate from './Certificate';
-
-export default Certificate;
